[MOV] website_sale: removed the website_sale_link_tracker module and moved it to the website_sale.
Task Id #1929272
