+++ addons/website_sale/controllers/backend.py
@@ -29,6 +29,7 @@ class WebsiteSaleBackend(WebsiteBackend):
         )
         results['dashboards']['sales'] = sales_values
 
+        results['dashboards']['sales']['utm_graph'] = self.fetch_utm_data(date_from, date_to)
         results['groups']['sale_salesman'] = request.env['res.users'].has_group('sales_team.group_sale_salesman')
         if not results['groups']['sale_salesman']:
             return results
@@ -127,6 +128,41 @@ class WebsiteSaleBackend(WebsiteBackend):
         }]
 
         return results
+
+    def fetch_utm_data(self, date_from, date_to):
+        sale_utm_domain = [
+            ('website_id', '!=', False),
+            ('state', 'in', ['sale', 'done']),
+            ('confirmation_date', '>=', date_from),
+            ('confirmation_date', '<=', date_to)
+        ]
+
+        orders_data_groupby_campaign_id = request.env['sale.order'].read_group(
+            domain=sale_utm_domain + [('campaign_id', '!=', False)],
+            fields=['amount_total', 'id', 'campaign_id'],
+            groupby='campaign_id')
+
+        orders_data_groupby_medium_id = request.env['sale.order'].read_group(
+            domain=sale_utm_domain + [('medium_id', '!=', False)],
+            fields=['amount_total', 'id', 'medium_id'],
+            groupby='medium_id')
+
+        orders_data_groupby_source_id = request.env['sale.order'].read_group(
+            domain=sale_utm_domain + [('source_id', '!=', False)],
+            fields=['amount_total', 'id', 'source_id'],
+            groupby='source_id')
+
+        return {
+            'campaign_id': self.compute_utm_graph_data('campaign_id', orders_data_groupby_campaign_id),
+            'medium_id': self.compute_utm_graph_data('medium_id', orders_data_groupby_medium_id),
+            'source_id': self.compute_utm_graph_data('source_id', orders_data_groupby_source_id),
+        }
+
+    def compute_utm_graph_data(self, utm_type, utm_graph_data):
+        return [{
+            'utm_type': data[utm_type][1],
+            'amount_total': data['amount_total']
+        } for data in utm_graph_data]
 
     def _compute_sale_graph(self, date_from, date_to, sales_domain, previous=False):
         days_between = (date_to - date_from).days
