+++ addons/website_sale/static/src/img/website_sale_chart_demo.png
