+++ addons/website_sale/static/src/scss/website_sale_dashboard.scss
@@ -69,4 +69,13 @@
         width: 100%;
         height: 30em;
     }
+    .o_dashboard_utms {
+        .utm_chart_image {
+            display: block;
+            margin: auto;
+        }
+        .utm_dropdown {
+            margin-top: -5px;
+        }
+    }
 }
