+++ addons/website_sale/static/src/xml/website_sale_dashboard.xml
@@ -117,4 +117,34 @@
             </div>
         </div>
     </t>
+
+    <t t-extend="website_sale.products_table">
+        <t t-jquery=".o_top_margin .row .col-12" t-operation="attributes">
+            <attribute name="class" value="col-lg-6 col-12" />
+        </t>
+        <t t-jquery=".o_top_margin .row" t-operation="append">
+            <div class="col-lg-6 col-12 o_dashboard_utms">
+                <div>
+                    <h4 class="float-left">REVENUE BY</h4>
+                    <t t-call="website_sale.LinkTrackersDropDown"/>
+                </div>
+                <div class="o_utm_no_data_img">
+                    <img src="website_sale/static/src/img/website_sale_chart_demo.png" alt="There isn't any UTM tag detected in orders" class="utm_chart_image image-responsive mt8"/>
+                </div>
+                <div class="o_utm_data_graph"/>
+            </div>
+        </t>
+    </t>
+
+    <t t-name="website_sale.LinkTrackersDropDown">
+        <div class="dropdown">
+            <button class="btn btn-secondary dropdown-toggle utm_dropdown ml4" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true"><span class="utm_button_name">Campaigns</span>
+            </button>
+            <div class="dropdown-menu" role="menu" aria-labelledby="utm_dropdown">
+                <a name="campaign_id" class="dropdown-item js_utm_selector" role="menuitem">Campaigns</a>
+                <a name="medium_id" class="dropdown-item js_utm_selector" role="menuitem">Medium</a>
+                <a name="source_id" class="dropdown-item js_utm_selector" role="menuitem">Sources</a>
+            </div>
+        </div>
+    </t>
 </templates>
