--- addons/website_sale_link_tracker/data/sale_order_demo.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-    <record id="utm_source_sale_order_0" model="utm.source">
-        <field name="name">Sale Promotion 1</field>
-    </record>
-    <record id="website_sale.website_sale_order_1" model="sale.order">
-        <field name="campaign_id" ref="utm.utm_campaign_email_campaign_products"/>
-        <field name="source_id" ref="website_sale_link_tracker.utm_source_sale_order_0"/>
-        <field name="medium_id" ref="utm.utm_medium_email"/>
-    </record>
-    <record id="website_sale.website_sale_order_2" model="sale.order">
-        <field name="campaign_id" ref="utm.utm_campaign_email_campaign_products"/>
-        <field name="source_id" ref="website_sale_link_tracker.utm_source_sale_order_0"/>
-        <field name="medium_id" ref="utm.utm_medium_email"/>
-    </record>
-    <record id="website_sale.website_sale_order_3" model="sale.order">
-        <field name="campaign_id" ref="utm.utm_campaign_email_campaign_products"/>
-        <field name="source_id" ref="website_sale_link_tracker.utm_source_sale_order_0"/>
-        <field name="medium_id" ref="utm.utm_medium_email"/>
-    </record>
-    <record id="website_sale.website_sale_order_4" model="sale.order">
-        <field name="campaign_id" ref="utm.utm_campaign_email_campaign_products"/>
-        <field name="source_id" ref="website_sale_link_tracker.utm_source_sale_order_0"/>
-        <field name="medium_id" ref="utm.utm_medium_email"/>
-    </record>
-</odoo>
