--- addons/website_sale_link_tracker/i18n/cs.po
@@ -1,95 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# * website_sale_link_tracker
-# 
-# Translators:
-# Martin Trigaux, 2018
-# trendspotter <j.podhorecky@volny.cz>, 2018
-# 
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-10-08 06:49+0000\n"
-"PO-Revision-Date: 2018-08-24 09:35+0000\n"
-"Last-Translator: trendspotter <j.podhorecky@volny.cz>, 2018\n"
-"Language-Team: Czech (https://www.transifex.com/odoo/teams/41243/cs/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: cs\n"
-"Plural-Forms: nplurals=4; plural=(n == 1 && n % 1 == 0) ? 0 : (n >= 2 && n <= 4 && n % 1 == 0) ? 1: (n % 1 != 0 ) ? 2 : 3;\n"
-
-#. module: website_sale_link_tracker
-#: model:ir.model.fields,field_description:website_sale_link_tracker.field_sale_report__campaign_id
-msgid "Campaign"
-msgstr "Kampaň"
-
-#. module: website_sale_link_tracker
-#. openerp-web
-#: code:addons/website_sale_link_tracker/static/src/xml/website_sale_dashboard_utm_chart.xml:23
-#: code:addons/website_sale_link_tracker/static/src/xml/website_sale_dashboard_utm_chart.xml:26
-#, python-format
-msgid "Campaigns"
-msgstr "Kampaně"
-
-#. module: website_sale_link_tracker
-#: model_terms:ir.ui.view,arch_db:website_sale_link_tracker.sale_order_from_website_link_tracker
-msgid "Marketing"
-msgstr "Marketing"
-
-#. module: website_sale_link_tracker
-#. openerp-web
-#: code:addons/website_sale_link_tracker/static/src/xml/website_sale_dashboard_utm_chart.xml:27
-#: model:ir.model.fields,field_description:website_sale_link_tracker.field_sale_report__medium_id
-#, python-format
-msgid "Medium"
-msgstr "Střední"
-
-#. module: website_sale_link_tracker
-#. openerp-web
-#: code:addons/website_sale_link_tracker/static/src/xml/website_sale_dashboard_utm_chart.xml:10
-#, python-format
-msgid "REVENUE BY"
-msgstr ""
-
-#. module: website_sale_link_tracker
-#: model:utm.source,name:website_sale_link_tracker.utm_source_sale_order_0
-msgid "Sale Newsletter"
-msgstr "Prodejní zpravodaj"
-
-#. module: website_sale_link_tracker
-#: model:ir.model,name:website_sale_link_tracker.model_sale_order
-msgid "Sale Order"
-msgstr "Prodejní objednávka"
-
-#. module: website_sale_link_tracker
-#: model:ir.model,name:website_sale_link_tracker.model_sale_report
-msgid "Sales Analysis Report"
-msgstr "Zpráva analýzy prodeje"
-
-#. module: website_sale_link_tracker
-#: model:ir.model.fields,field_description:website_sale_link_tracker.field_sale_report__source_id
-msgid "Source"
-msgstr "Zdroj"
-
-#. module: website_sale_link_tracker
-#. openerp-web
-#: code:addons/website_sale_link_tracker/static/src/xml/website_sale_dashboard_utm_chart.xml:28
-#, python-format
-msgid "Sources"
-msgstr "Zdroje"
-
-#. module: website_sale_link_tracker
-#. openerp-web
-#: code:addons/website_sale_link_tracker/static/src/xml/website_sale_dashboard_utm_chart.xml:14
-#, python-format
-msgid "There isn't any UTM tag detected in orders"
-msgstr ""
-
-#. module: website_sale_link_tracker
-#. openerp-web
-#: code:addons/website_sale_link_tracker/static/src/xml/website_sale_dashboard_utm_chart.xml:5
-#, python-format
-msgid "col-lg-6 col-12"
-msgstr ""
