--- addons/website_sale_link_tracker/models/sale_order.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from odoo import models
-
-
-class SaleOrder(models.Model):
-    _name = "sale.order"
-    _inherit = ['utm.mixin', 'sale.order']
