--- addons/website_sale_link_tracker/static/src/scss/website_sale_link_tracker.scss
@@ -1,10 +1,0 @@
-.o_dashboard_utms {
-    .utm_chart_image {
-        display: block;
-        margin: auto;
-    }
-
-    .utm_dropdown {
-        margin-top: -5px;
-    }
-}
