--- addons/website_sale_link_tracker/views/assets.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <template id="assets_backend" inherit_id="web.assets_backend">
-        <xpath expr="." position="inside">
-            <script type="text/javascript" src="/website_sale_link_tracker/static/src/js/website_sale_dashborad_utm_chart.js"></script>
-            <link rel="stylesheet" type="text/scss" href="/website_sale_link_tracker/static/src/scss/website_sale_link_tracker.scss"/>
-        </xpath>
-    </template>
-</odoo>
